requirements.txt
Flask==3.0.3
requests
imaplib2
SQLAlchemy
psycopg2-binary
python-dotenv
